chore(fsrs): move legacy exports
