+++ .changeset/move-legacy-exports.md
@@ -1,0 +1,7 @@
+---
+"ts-fsrs": major
+---
+
+chore: move legacy scheduler models, types, constants, and helpers behind the
+legacy barrel while re-exporting SRS primitives and schemas from srs-kit and
+the FSRS memory-state schema.
