+++ packages/fsrs/bench/monotonic-interval.bench.ts
@@ -1,6 +1,6 @@
 import { defineScheduler, Rating } from '@open-spaced-repetition/srs-kit'
 import { dateChrono } from '@open-spaced-repetition/srs-kit/chrono/date'
-import { createEmptyCard } from 'ts-fsrs/default'
+import { createEmptyCard } from 'ts-fsrs/legacy/default'
 import { fsrs } from 'ts-fsrs/legacy/fsrs'
 import {
   calculateScheduleDay,
