+++ packages/fsrs/src/kit/types.ts
@@ -1,6 +1,10 @@
 import type { ModelCore } from '@open-spaced-repetition/srs-kit/model'
 import type { FSRS6Algorithm } from '../models/fsrs-6/algorithm.js'
-import type { FSRSState } from '../models.js'
+
+export interface FSRSState {
+  stability: number
+  difficulty: number
+}
 
 /** @internal */
 export type IFSRSModel = ModelCore<{
