+++ packages/fsrs/src/legacy/impl/basic_scheduler.ts
@@ -1,14 +1,6 @@
-import { TypeConvert } from '@/convert.js'
-import { date_scheduler } from '@/help.js'
 import type { IFSRSModel } from '@/kit/index.js'
-import { withFuzzing } from '@/middlewares/fuzzing/core.js'
-import { calculateLearningSteps } from '@/middlewares/learning-steps/core.js'
-import type {
-  LearningStepsConfig,
-  LearningStepsResolver,
-  LearningStepsResult,
-} from '@/middlewares/learning-steps/types.js'
-import { calculateScheduleDays } from '@/middlewares/monotonic-interval/core.js'
+import { TypeConvert } from '@/legacy/convert.js'
+import { date_scheduler } from '@/legacy/help.js'
 import {
   type Card,
   type CardInput,
@@ -18,8 +10,16 @@ import {
   Rating,
   type RecordLogItem,
   State,
-} from '@/models.js'
-import type { int } from '@/types.js'
+} from '@/legacy/models.js'
+import type { int } from '@/legacy/types.js'
+import { withFuzzing } from '@/middlewares/fuzzing/core.js'
+import { calculateLearningSteps } from '@/middlewares/learning-steps/core.js'
+import type {
+  LearningStepsConfig,
+  LearningStepsResolver,
+  LearningStepsResult,
+} from '@/middlewares/learning-steps/types.js'
+import { calculateScheduleDays } from '@/middlewares/monotonic-interval/core.js'
 import { AbstractScheduler } from '../abstract_scheduler.js'
 import { StrategyMode, type TStrategyHandler } from '../strategies/index.js'
