+++ packages/fsrs/src/legacy/index.ts
@@ -1,0 +1,26 @@
+export * from './abstract_scheduler.js'
+export * from './constant.js'
+export * from './convert.js'
+export * from './default.js'
+export * from './fsrs.js'
+export * from './help.js'
+export * from './impl/basic_scheduler.js'
+export * from './impl/long_term_scheduler.js'
+export type {
+  Card,
+  CardInput,
+  DateInput,
+  FSRSHistory,
+  FSRSParameters,
+  FSRSReview,
+  GradeType,
+  RatingType,
+  RecordLog,
+  RecordLogItem,
+  ReviewLog,
+  ReviewLogInput,
+  StateType,
+  Steps,
+} from './models.js'
+export * from './strategies/index.js'
+export type * from './types.js'
