+++ packages/fsrs/src/models/fsrs-4dot5/algorithm.spec.ts
@@ -1,6 +1,6 @@
-import { grades } from '@open-spaced-repetition/srs-kit'
+import { type Grade, grades, Rating } from '@open-spaced-repetition/srs-kit'
 import { describe, expect, it } from 'vitest'
-import { type FSRSState, type Grade, Rating } from '@/models.js'
+import type { FSRSState } from '@/kit/types.js'
 import {
   FSRS4Dot5_DECAY,
   FSRS4Dot5_DEFAULT_WEIGHTS,
