+++ packages/fsrs/src/models/fsrs-5/algorithm.spec.ts
@@ -1,5 +1,5 @@
+import { type Grade, Rating } from '@open-spaced-repetition/srs-kit'
 import { describe, expect, it } from 'vitest'
-import { type Grade, Rating } from '@/models.js'
 import { FSRS5Algorithm, forgetting_curve } from './algorithm.js'
 import {
   FSRS5_DECAY,
