+++ packages/fsrs/src/models/fsrs-5/constants.ts
@@ -1,5 +1,5 @@
 import type { ModelBounds } from '@open-spaced-repetition/srs-kit/model'
-import type { FSRSState } from '@/models.js'
+import type { FSRSState } from '@/kit/types.js'
 
 export const FSRS5_DECAY = 0.5
 export const FSRS5_FACTOR = 19 / 81
